Modify DEP 34 language tests in preparation for adding analyzer support.
Both of these tests used "VMOptions" to enable the new behavior.  They
need to use "SharedOptions" so that the new behavior is also enabled
in dart2js and analyzer.

Also, since mixin_super_bound2_test verifies that the appropriate
errors are thrown at runtime when type bounds are incorrect, it's
going to have a static warning when run through the analyzer (once the
analyzer supports DEP 34).  In order to prevent this from leading to a
buildbot failure, we need to change the test into a multitest:

- mixin_super_bound2_test/01 reproduces the previous behavior of the
  test, throwing the appropriate errors at runtime and generating a
  static warning when analyzed.

- mixin_super_bound2_test/none lacks the incorrect type bounds, so it
  doesn't produce and runtime errors or static warnings.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org//1254933003 .