dart2js CPS: Fix an incorrect assertion.
When computing captured variables in constructors, there was an assert
that an assignment in one of the initializers must be a parameter
assignment (since those are the only locals in scope for the
initializers).

However, the initializers can contain closures, which can have
assignments to their own local variables so the assert was not
correct.

BUG=
R=karlklose@google.com

Review URL: https://codereview.chromium.org//1278463005 .