sdk/dart now detects an available build
The shell scripts in sdk/sdk are using ReleaseIA32 as the default build
configuration in the case one is not provided as an environment
variable. This is arbitrary so a check of the available builds is
performed and if ReleaseIA32 is available it is selected. Otherwise, an
arbitrary configuration is selected (currently the last one
lexicographically). In the case where there are no subfolders in the
"out" folder, ReleaseIA32 is selected, in order to produce the same
error that used to be produced before (this can be changed to a custom
message for user-friendliness).

Currently only the "dart" shell script is changed and if the change is
accepted, the others will be changed too.

R=floitsch@google.com, zerny@google.com

Review URL: https://codereview.chromium.org//1276333002 .