dart2js cps: Bugfixes in .push rewrites.
Two consecutive pushes would sometimes be merged in a way that one of
the arguments could fall out of scope.

List.addAll with a literal list as argument would sometimes be rewritten
to a push, even though the list could be modified before the use.

BUG=
R=kmillikin@google.com

Review URL: https://codereview.chromium.org//1286123005 .