VM: Materialize stub pc-markers (smi 0) eagerly on deoptimization.
Deferred objects are malloced, and there is no need to allocate one
for the stub pc-marker (smi zero).

This is also part of preparing for running precompiled stub code: when
changing the pc-marker to hold the stub's code object, we want to materialize
it eagerly, so that the stack frame is valid after DeoptimizeFillFrame.

BUG=
R=rmacnak@google.com

Review URL: https://codereview.chromium.org//1289243002 .