Support older SDKs in analyzer.
In Dart 1.12, the location of the "libraries.dart" file, which
analyzer uses to find the contents of the SDK, will be changed.
However, 1.12 hasn't been released yet, so we need analyzer to
continue supporting the current stable release of the SDK.

This change causes analyzer to first look in the new location for
"libraries.dart", and failing that, fall back to the old (pre-1.12)
location.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org//1296033004 .