Band-aid fix for issue #24191.
This is sufficient to avoid throwing an exception in the resolver,
which in turn will prevent the analysis server from going into an
infinite analysis loop.

I'll work on a more complete fix (which infers the proper type in all
cases) for the next release.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org//1307353002 .