dart2js cps: Use a global counter to generate hash codes.
This gives almost 30x compile-time speedup on ImagingGaussianBlurOnce.

Based on profiling with Observatory, it seems to be a performance bug
in how hash codes are stored on objects. The majority of time is spent
in Object_GetHash, which does not compute a hash, but looks it up in a
weak hash table in the VM.

R=kmillikin@google.com

Review URL: https://codereview.chromium.org//1311673009 .