dart2js cps: Generate interceptors at use-site and share afterwards.
Previously, interceptors were generated at the definition-site to ensure
sharing, and sometimes sunk to its use-site if desirable.

Now they are (again) generated at use-site, then pulled out of loops,
and replaced by dominating calls to getInterceptor.

This means interceptors are only computed if they really are needed,
but on the other hand, they are sometimes computed more than once, when
there are uses that don't have a common dominator.

Interceptor constants are currently not shared/hoisted as it is not
clear when this is beneficial.

R=kmillikin@google.com

Review URL: https://codereview.chromium.org//1313783003 .