Revert "This is a fixed up version of:"
This reverts commit dd7767d95eda87054844e0b24ebd2ce9b04354bb.

This made the analyzer fail:
http://build.chromium.org/p/client.dart/builders/analyzer_experimental-linux-release-be/builds/3418/

Please note that this is just not the new test (which I would just have
suppressed)
It seems that in --checked --host-checked mode we see a range of new
failures, because of the change to flags.

It also made dart2js fail on the new tests

Ivan: Will you look into this with the analyzer people?

R=iposva@google.com
BUG=

Review URL: https://codereview.chromium.org//1308743006 .