dart2js cps: Track existence of try statements in resolution instead of traversing AST/IR nodes to check for it.
This change removes one traversal in the backend and it is one step in the direction of recording more inforamtion in resolution; the compuation of DartCapturedVariables, for example, could also be done during resolution and stored in the TreeElements.

The disadvantage is that resolution may look at more code than TypePropagation, but in this specific case I do not think that we are realisticly able to eliminate try statements in functions that are small enough to be inlined.

R=kmillikin@google.com

Review URL: https://codereview.chromium.org//1311403003 .