Add resource metadata for open files.
Start tracking the read/write information for files in the new
resource management class.

There is currently no time information, I will refactor the timings so that the
static stopwatch from the socket_patch class is moved into the resource class
instead in a follow up (keeping this simply and only about the file addition).

The testing setup may seem a bit strange, but it is the only way I
seem to be able to guarantee that setup actually finishes and that we
get it cleaned up. John may know of a better way

R=sgjesse@google.com, johnmccutchan@google.com
BUG=

Review URL: https://codereview.chromium.org//1326703004 .