Make type system changes backwards compatible
This CL eliminates the additional typeSystem arguments added to the API by https://codereview.chromium.org/1329743005/  wherever possible, instead projecting out the typeSystem from the context as available.  Where not possible, the additional argument is made optional, defaulting to the standard TypeSystemImpl.  I believe that the result of this should be backwards compatible for clients (but not implementors) of the API.

BUG=
R=brianwilkerson@google.com, paulberry@google.com

Review URL: https://codereview.chromium.org//1355773002 .