Do not emit type metadata for classes that are only emitted for rti.
Classes only emitted for rti are empty shells and, even if they are
marked as reflectable, are not otherwise referenced from the program. So
we should not try to emit type metatdata because it is neither needed
for is it always complete (as most of the class, including its type
variables, is missing).

Fixes #24493.

BUG= http://dartbug.com/24493
R=sigmund@google.com

Review URL: https://codereview.chromium.org/1394483002 .