Use interceptors for is-checks (version 2).
Use the $isXXX properties for type tests on raw (non-parameterized) types.

Use special builtin operations for the weird JSArray type hierarchy.

Still TODO:

If the only use of an interceptor is a single 'getInterceptor(value).$isXXX', and XXX is not a supertype of a native type, replace with 'value instanceof XXX'.

R=kmillikin@google.com

Review URL: https://codereview.chromium.org/1385423002 .