VM: Use stm/ldm when saving/restoring live registers on ARM.
The order in which live registers are pushed/restored in
the slow-path code is reversed so that it matches the order
of the ARM stm/ldm instructions.

The same could be done for FPU registers. It would only affect
optimized code though. In precompiled and unoptimized code, there
are no live FPU registers at slow-path instructions.

BUG=
R=srdjan@google.com

Review URL: https://codereview.chromium.org/1387303003 .