Create a breadcrumb trail for analyzing of task model extensions.
This change paves the way for modifying
pkg/analyzer/tool/task_dependency_graph.dart so that it will be able to
understand the extensibility mechanisms in the task model.
Specifically, it adds the ExtensionPointId annotation, which
task_dependency_graph.dart will be able to use to locate the
TaskDescriptors that are plugged into each task model extension point.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/1401543002 .