Fixed improper construction of custom elements.
Was a too aggressive on upgrading a custom element.  If a JS object is a custom element never construct using the actual custom class in wrap_jso only allocate the TemplateElement or HtmlElement and let the upgrader or createdCallback do the upgrading.  This was causing timing issues with nested custom elements that were data bound.

TBR=alanknight@google.com

Review URL: https://codereview.chromium.org/1417363003 .