dart2js cps: More interceptor optimizations and fixes.
When the result of getInterceptor is either null or a known constant,
it is compiled to `x && const` or `x == null ? x : const` depending
on whether there are any other falsy values for that type.

This required some changes to share_interceptors so it can share/hoist
the constant subexpression separately from the whole expression.

The Interceptor root class is removed from the intercepted classes of
getInterceptor when it is "shadowed" by all the other intercepted classes.
E.g. if the input is a number or a string, there is no need
to intercept for {JSNumber, JSString, Interceptor}, because JSNumber
and JSString collectively shadow Interceptor.

R=sra@google.com

Review URL: https://codereview.chromium.org/1409803003 .