move TypeSystem classes to their own file also sorted members per analyzer style (?)
figured it would be easier to work on as a 0.5k file vs ~15k resolver.dart :)

For now I left TypeProvider where it was. Not sure if it should move here or into its own file as well.

For backwards compat, all of the types are re-exported from resolver.dart.

R=brianwilkerson@google.com

Review URL: https://codereview.chromium.org/1407923009 .