dart2js cps: Better side-effect tracking in loops.
This fixes an issue in the bounds checker that made it think a nested
loop had side effects.

In the first pass, the bounds checker would conservatively assume
each loop had side effects, but then mark it as effect-free for the
next pass if it did not see any effects in the loop. For nested loops
however, the "conservative" side effect of the inner loop would prevent
the outer loop from being marked as effect-free.

A separate pass now analysis loops for side effects.

BUG=
R=kmillikin@google.com

Review URL: https://codereview.chromium.org/1426633005 .