Mark returns from inside the try of a try/catch as ones that exit the try.
At construction time, jumps from within the try block of a try/catch are
marked as ones that exit the try block.  This is relied on for correctness.
Specifically, it prevents moving code from outside the try into the scope of
the try.

Return jumps were from within the try were not marked because it was
unnecessary.  In the presence of inlining, returns are turned into jumps to a
local join-point continuation.  These jumps must be marked for correctness.

R=asgerf@google.com
BUG=

Review URL: https://codereview.chromium.org/1414043011 .