dart2js cps: Propagate container types for lists.
Container type masks depend on the allocation site of an object, but the
allocation site "key" is an AST node. We need to assign types to nodes
in the type propagation pass where the AST nodes are (and should be)
inaccessible, so we cannot directly query the type inference from there.

The key could be changed to an opaque "AllocationSiteID" object with no
members (just a key), but for now, we're just preserving the type mask
for container types from build-time.

Also has some other type propagation improvements related to containers.

BUG=
R=kmillikin@google.com

Review URL: https://codereview.chromium.org/1416723008 .