js_ast: Avoid precedence issue with negative numbers as receivers at calls
This change makes LiteralNumber and DeferredNumber have a precedence based on the value so that the sign always binds to the number.

An alternative is to change constant generation to built the negation of a positive constant, but this approach does not work for deferred numbers since the sign is unknown at AST generation time.

R=sigmund@google.com

Review URL: https://codereview.chromium.org/1438633002 .