dart2js cps: Clean up and avoid processing unreachable code.
Primitives now have a getter 'hasValue' to distinguish primitives that
don't return a value, such as SetField.

The interceptor for TypeTestViaFlag is now inserted on-demand.

Non-instantiated intercepted classes are removed the set of
intercepted classes. This avoids an assertion failure, although it
would be better to avoid adding them in the first place.

The type of closure fields for boxes and type variables no longer go
through the global type inference.

A test fails because the type inference infers an empty type for a
closure field for a torn-off `.call` method. It's an existing problem
that we now have at least one failing test for.

CreateFunction has been removed from CPS and Tree.

BUG=
R=sra@google.com

Review URL: https://codereview.chromium.org/1474713002 .